user can still play soundcapsule over again *fix
