+++ client/src/pages/capsules/SoundCapsule/FinishedSoundCapsule.jsx
@@ -2,11 +2,23 @@ import React, { useEffect, useState, useContext } from "react";
 import { Link, useNavigate } from "react-router-dom";
 import note from "../../../assets/images/soundcapsule/note2.svg";
 import singleNote from "../../../assets/images/soundcapsule/note3.svg";
-
-const FinishedSoundCapsule = ({ update }) => {
+import { User } from "../../../application";
+const FinishedSoundCapsule = ({ update, id }) => {
+  const { user, setUser } = useContext(User);
   useEffect(async () => {
     await update();
   }, []);
+
+  const capsuleNameFromDatabase = user.finishedCapsules.map((capsuleName) => {
+    return capsuleName.name;
+  });
+
+  const filteredCapsuleNamesFromUserDatabase = capsuleNameFromDatabase.find(
+    (capsuleName) => {
+      console.log(capsuleName);
+      return capsuleName === id;
+    }
+  );
 
   return (
     <div className="d-flex justify-content-around align-items-center flex-column vh-100 bg-capsule">
@@ -62,13 +74,25 @@ const FinishedSoundCapsule = ({ update }) => {
       >
         <img width={50} src={singleNote} alt="Some note" />
       </div>
-      <div className="text-center">
-        <h3 className="text-capsule fw-bold">Fullført lydkapselen</h3>
-      </div>
+      {filteredCapsuleNamesFromUserDatabase ? (
+        <>
+          <div className="text-center">
+            <h3 className="text-capsule fw-bold">
+              Du har allerede fullført denne lydkapselen
+            </h3>
+          </div>
+        </>
+      ) : (
+        <>
+          <div className="text-center">
+            <h3 className="text-capsule fw-bold">Fullført lydkapselen</h3>
+          </div>
 
-      <div>
-        <h4>+20 poeng</h4>
-      </div>
+          <div>
+            <h4>+20 poeng</h4>
+          </div>
+        </>
+      )}
 
       <div className="text-center">
         <Link to="/map">
